fix: Fix Handling CldUploadWidget Callback Updates to State (#138)
# Description

The CldUploadWidget is facing an issue where the state setter isn't
properly updating after route changes. I can't quite pin down the root
problem, but my thinking is that due to the route changes, a reference
is lost, and it's unable to work even though the callbacks seem to fire
appropriately.

The solution proposed here is to take advantage of local state in the
UploadWidget component to track the callback results and then bubble
those up to the event handlers, including `onUpload` and newly added
`onError`.

This is a breaking change, as `onUpload` will now remove the `error`
argument as the first argument, given the `error` is tracked separately.

```
  <CldUploadWidget
        uploadPreset="next-cloudinary-unsigned"
        onUpload={(result, widget) => {
           // success
        }}
        onError={(result, widget) => {
           // error
        }}
      >
```

Questions:
* Would it make sense to store a single root state object with both the
error and the result and triggering the handler logic based off of any
changes to that?
* Are there any negative consequences for tracking state this way? Only
issue I can think of to start is user will need to add `use client` as
it's now using client code

## Issue Ticket Number

Fixes #128 

<!-- Specify above which issue this fixes by referencing the issue
number (`#<ISSUE_NUMBER>`) or issue URL. -->
<!-- Example: Fixes
https://github.com/colbyfayock/next-cloudinary/issues/<ISSUE_NUMBER> -->

## Type of change

<!-- Please select all options that are applicable. -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Fix or improve the documentation
- [ ] This change requires a documentation update


# Checklist

<!-- These must all be followed and checked. -->

- [ ] I have followed the contributing guidelines of this project as
mentioned in [CONTRIBUTING.md](/CONTRIBUTING.md)
- [ ] I have created an
[issue](https://github.com/colbyfayock/next-cloudinary/issues) ticket
for this PR
- [ ] I have checked to ensure there aren't other open [Pull
Requests](https://github.com/colbyfayock/next-cloudinary/pulls) for the
same update/change?
- [ ] I have performed a self-review of my own code
- [ ] I have run tests locally to ensure they all pass
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes needed to the documentation

BREAKING CHANGE: CldUploadWidget onUpload property no longer passes an error object as the first argument. The first argument is now the result object. A separate onError property was added to handle errors.